requirements.txt
APScheduler==3.10.4
blinker==1.8.1
click==8.1.7
Flask==3.0.3
Flask-HTTPAuth==4.8.0
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.2